requirements.txt
asgiref==3.5.2
black==22.10.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==3.0.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
defusedxml==0.7.1
Django==4.1.3
django-templated-mail==1.1.1
django-filter~=22.1
django-phonenumber-field==7.0.0
djoser==2.1.0
drf-nested-routers==0.93.4
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
oauthlib==3.2.2
packaging==21.3
pathspec==0.10.2
phonenumbers==8.13.0
Pillow==9.3.0
platformdirs==2.5.4
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.12
djangorestframework~=3.14.0
